build-test-repo: create a git bundle with some pathological cases
In addition to the tree, parent, author and committer headers,
a git commit can also contain encoding, gpgsig and mergetag.

This script creates repository with all 8 combinations of
encoding, gpgsig and mergetag and creates a bundle from it,
to make it easy to use in tests.

Although the commit dates and authors are hard-coded, this does not
make the repository stays the same through re-runs, because gpg also
uses the current date when signing stuff.